+++ src/argus/incident/views.py
@@ -1,5 +1,7 @@
+import logging
 import secrets
 
+from django.conf import settings
 from django.db import IntegrityError
 
 from django_filters import rest_framework as filters
@@ -18,7 +20,9 @@ from rest_framework.reverse import reverse
 
 from argus.auth.models import User
 from argus.drf.permissions import IsSuperuserOrReadOnly
+from argus.incident.ticket.base import TicketPluginException
 from argus.util.datetime_utils import INFINITY_REPR
+from argus.util.utils import import_class_from_dotted_path
 
 from .forms import AddSourceSystemForm
 from .filters import IncidentFilter, SourceLockedIncidentFilter
@@ -32,6 +36,7 @@ from .models import (
 )
 from .serializers import (
     UpdateAcknowledgementSerializer,
+    EmptySerializer,
     EventSerializer,
     IncidentPureDeserializer,
     IncidentSerializer,
@@ -47,6 +52,8 @@ from .serializers import (
     TagSerializer,
     IncidentTagRelation,
 )
+
+LOG = logging.getLogger(__name__)
 
 
 # Used in OpenApiParameter
@@ -250,6 +257,64 @@ class IncidentViewSet(
             return Response(serializer.data)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@extend_schema_view(
+    update=extend_schema(
+        request=EmptySerializer,
+    ),
+)
+class TicketPluginViewSet(viewsets.ViewSet):
+    permission_classes = [IsAuthenticated]
+    serializer_class = IncidentTicketUrlSerializer
+    queryset = Incident.objects.all()
+
+    def update(self, request, incident_pk=None):
+        incident = get_object_or_404(self.queryset, pk=incident_pk)
+
+        if incident.ticket_url:
+            serializer = self.serializer_class(data={"ticket_url": incident.ticket_url})
+            if serializer.is_valid():
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        plugin = getattr(settings, "TICKET_PLUGIN", None)
+
+        if not plugin:
+            return Response(
+                data="No path to ticket plugin can be found in the settings. Please update the setting 'TICKET_PLUGIN'.",
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        try:
+            ticket_class = import_class_from_dotted_path(plugin)
+        except Exception:
+            LOG.exception("Could not import ticket plugin from path %s", plugin)
+            return Response(
+                data="Ticket plugins are incorrectly configured.",
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        try:
+            url = ticket_class.create_ticket(incident)
+        except TicketPluginException as e:
+            return Response(
+                data=str(e),
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        if url:
+            incident.ticket_url = url
+            incident.save(update_fields=["ticket_url"])
+            serializer = self.serializer_class(data={"ticket_url": incident.ticket_url})
+            if serializer.is_valid():
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        return Response(
+            data="No url could be generated. Please check that the ticket plugin provides a function to create tickets.",
+            status=status.HTTP_400_BAD_REQUEST,
+        )
 
 
 class IncidentTagViewSet(