+++ tests/incident/test_views.py
@@ -654,6 +654,13 @@ class IncidentViewSetTestCase(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(Incident.objects.get(id=incident_pk).ticket_url, data["ticket_url"])
 
+    def test_can_get_existing_ticket_url_of_incident(self):
+        ticket_url = "www.example.com"
+        pk = StatefulIncidentFactory(ticket_url=ticket_url).pk
+        response = self.client.put(path=f"/api/v2/incidents/{pk}/ticket/")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["ticket_url"], ticket_url)
+
     def test_can_get_my_incidents(self):
         incident_pk = self.add_open_incident_with_start_event_and_tag().pk
         other_incident_pk = StatefulIncidentFactory().pk